refactor(states): return err on getting or setting a state
We somehow need to handle errors in the case if a storage implementation
(e.g. redis) can't get or set the state because for example the redis
instance has been killed, there's no internet connection if the redis
instance not on local host etc.

But we even can't get those errors because "GetState" and "SetState"
methods of a storage don't return errors.

In the manager and filter I just return those storage errors, but I
think we can make something similar to telebot's "bot.OnError" function,
where a developer can for example send a message to the user with the
text "Something goes wrong, click on /start" if we can't get the state
for finding the right handler.